# wtu-server-core/src/main/java/es/coffeebyt/wtu/api/ApiError.java
package es.coffeebyt.wtu.api;

import lombok.AllArgsConstructor;
import lombok.Data;

@Data
@AllArgsConstructor
public class ApiError {

    private String timestamp;
    private int status;
    private String error;
    private String message;
    private String path;

}